requirements.txt
matplotlib==3.5.2
networkx==2.8.7
pandas==1.4.4
Pillow==9.3.0
torch==1.12.1
torchvision==0.13.1
